Remove the choice for MQTT logging if it is disabled
This is a second pass at https://github.com/esphome/issues/issues/6046

I have a situation where sometimes I have MQTT logging enabled, but
sometimes not depending on the device.

This PR removes the choice so that one doesn't have to remember that
MQTT logging is disabled, so that `esphome logs` goes right for the API
method (if available).